requirements.txt
aiogram==3.7.0
sqlalchemy[asyncio]
aiosqlite
asyncpg
python-dotenv
aiohttp
alembic
pytest
pytest-asyncio
